% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_projects.R
\name{tabulate_projects}
\alias{tabulate_projects}
\title{produce a quick table of projects and annotations}
\usage{
tabulate_projects(url = "https://itcr.cancer.gov/informatics-tools-table")
}
\arguments{
\item{url}{character(1) itcr.cancer.gov URL for tabular form of project descriptions}
}
\value{
a data.frame with 3 variables: title, description, and tag
}
\description{
produce a quick table of projects and annotations
}
\note{
Table was read on 28 April 2023.  A serialization from that date is in the data folder.
}
